requirements.txt
Adafruit-Blinka==8.69.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-ahtx0==1.0.28
adafruit-circuitpython-busdevice==5.2.15
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.86.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
click==8.3.1
Flask==3.1.2
gunicorn==24.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
lgpio==0.2.2.0
MarkupSafe==3.0.3
packaging==26.0
pyftdi==0.57.1
pyserial==3.5
python-dotenv==1.2.1
pyusb==1.3.1
redis==7.1.0
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
sysv_ipc==1.2.0
typing_extensions==4.15.0
Werkzeug==3.1.5
